+++ src/Services/Drivers/Plisio.php
@@ -92,6 +92,11 @@ class Plisio extends Driver
         return redirect($payment->failed_url);
     }
 
+    public static function secretKeys(): array
+    {
+        return ['secret_key', 'api_key'];
+    }
+
     public function integration(): array
     {
         return PaymentGateway::make('Plisio')