+++ src/Services/Drivers/Stripe.php
@@ -69,6 +69,11 @@ class Stripe extends Driver
         return redirect($payment->failed_url);
     }
 
+    public static function secretKeys(): array
+    {
+        return ['secret_key'];
+    }
+
     public function integration(): array
     {
         return PaymentGateway::make('StripeEmbedded')